refactor: extract Student Dashboard to StudentHome component
